# cmd/docs/swagger.yaml
basePath: /v1
definitions:
  models.Anime:
    properties:
      created_at:
        description: Timestamp when the anime was created
        example: "2025-02-20T19:27:00Z"
        type: string
      description:
        description: Description of the anime
        example: A story about ninjas.
        maxLength: 500
        type: string
      id:
        description: Unique identifier for the anime
        example: 1
        type: integer
      rating:
        description: Average rating of the anime
        example: 8.5
        maximum: 10
        minimum: 0
        type: number
      reviews:
        description: List of reviews for the anime (omitted unless necessary)
        items:
          $ref: '#/definitions/models.Review'
        type: array
      title:
        description: Title of the anime
        example: Naruto
        maxLength: 100
        minLength: 3
        type: string
      updated_at:
        description: Timestamp when the anime was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
    required:
    - title
    type: object
  models.AnimeCreateRequest:
    properties:
      description:
        description: Description of the anime
        example: A story about ninjas.
        maxLength: 500
        type: string
      rating:
        description: Rating of the anime
        example: 8.5
        maximum: 10
        minimum: 0
        type: number
      title:
        description: Title of the anime
        example: Naruto
        maxLength: 100
        minLength: 3
        type: string
    required:
    - title
    type: object
  models.AnimeResponse:
    properties:
      created_at:
        description: Timestamp when the anime was created
        example: "2025-02-20T19:27:00Z"
        type: string
      description:
        description: Description of the anime
        example: A story about ninjas.
        type: string
      id:
        description: Unique identifier for the anime
        example: 1
        type: integer
      rating:
        description: Average rating of the anime
        example: 8.5
        type: number
      title:
        description: Title of the anime
        example: Naruto
        type: string
      updated_at:
        description: Timestamp when the anime was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
    type: object
  models.Review:
    properties:
      animeId:
        description: ID of the anime being reviewed
        example: 1
        type: integer
      content:
        description: Content of the review
        example: This anime is amazing!
        maxLength: 500
        type: string
      created_at:
        description: Timestamp when the review was created
        example: "2025-02-20T19:27:00Z"
        type: string
      id:
        description: Unique identifier for the review
        example: 1
        type: integer
      rating:
        description: Rating given in the review (0-10)
        example: 9
        maximum: 10
        minimum: 0
        type: integer
      updated_at:
        description: Timestamp when the review was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
      userId:
        description: ID of the user who created the review
        example: 1
        type: integer
    required:
    - content
    - rating
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@myanimeapi.com
    name: API Support
  description: This is a sample API for managing anime and reviews.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: MyAnimeAPI
  version: "1.0"
paths:
  /v1/anime:
    get:
      description: Retrieve a list of all anime entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Anime'
            type: array
        "500":
          description: Failed to retrieve anime
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - "":
        - ""
      summary: Get all anime entries
      tags:
      - anime
    post:
      consumes:
      - application/json
      description: Create a new anime entry with the provided data
      parameters:
      - description: Anime data
        in: body
        name: anime
        required: true
        schema:
          $ref: '#/definitions/models.AnimeCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AnimeResponse'
        "400":
          description: Invalid input or missing required fields
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create anime
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new anime
      tags:
      - anime
  /v1/anime/{id}:
    delete:
      description: Delete an anime entry by its ID
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Anime not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete anime
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an anime
      tags:
      - anime
    get:
      description: Retrieve an anime by its ID
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Anime'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Anime not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve anime
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - "":
        - ""
      summary: Get an anime by ID
      tags:
      - anime
    put:
      consumes:
      - application/json
      description: Update an existing anime entry with the provided data
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated anime data
        in: body
        name: anime
        required: true
        schema:
          $ref: '#/definitions/models.Anime'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Anime'
        "400":
          description: Invalid input or ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Anime not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update anime
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an anime
      tags:
      - anime
  /v1/anime/{id}/reviews:
    get:
      description: Retrieve paginated reviews for an anime by its ID
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "400":
          description: Invalid ID format or pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve reviews
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - "":
        - ""
      summary: Get paginated reviews for an anime
      tags:
      - anime
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 'Use the format "Bearer <JWT_TOKEN>". Example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"